% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatin_state.R
\docType{data}
\name{chromatin_state}
\alias{chromatin_state}
\title{BED file contains position and name of chromatin state in mm10 genome}
\format{
A data frame with 566398 rows and 6 variables:
\describe{
\item{chr}{chromosome number}
\item{start}{start position, in bp}
\item{end}{end position, in bp}
\item{state}{state number }
\item{name}{sample name}
\item{state_name}{state name}
}
}
\usage{
chromatin_state
}
\description{
BED file contains position and name of chromatin state in mm10 genome
}
\keyword{datasets}
