% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromatinState.R
\docType{data}
\name{chromatinState}
\alias{chromatinState}
\title{chromatinState}
\format{
A GRanges object with 566398 ranges and 3 metadata columns:
\describe{
\item{chr}{chromosome number}
\item{start}{start position, in bp}
\item{end}{end position, in bp}
\item{state}{state number }
\item{name}{sample name}
\item{state_name}{state name}
}
}
\usage{
chromatinState
}
\description{
BED file contains position and name of chromatin state in mm10 genome
}
\keyword{datasets}
