% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorTable.R
\docType{data}
\name{colorTable}
\alias{colorTable}
\title{Weekly electricity consumption in France from 1996 to 2009 in MW
meteo and socio-economic variables related to it}
\format{
A list of data frame containing color associate at 3 ChromHMM models
chromatin state (15 / 18 or 25)
}
\source{
personal data
}
\usage{
colorTable
}
\description{
Weekly electricity consumption in France from 1996 to 2009 in MW
meteo and socio-economic variables related to it
}
\keyword{datasets}
