% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerAnnotation.R
\name{enhancerAnnotation}
\alias{enhancerAnnotation}
\title{Function which associated at each enhancer all gene in the interval arount the enhancer}
\usage{
enhancerAnnotation(enhancerTable, genome, interval = 5e+05, nCore = 1)
}
\arguments{
\item{enhancerTable}{GRanges table contains genomic position for each enhancer (ex. return by ChromHMM tool}

\item{genome}{a bed genome annotation file}

\item{interval}{a numeric value corresponding to the distance from enhancer where gene is looking for (default interval = 500000 (500kb))}

\item{nCore}{a numeric value corresponding to the number of core used (default nCore = 1)}
}
\value{
a table with infromation about gene associated with enhancer
}
\description{
Function which associated at each enhancer all gene in the interval arount the enhancer
}
