% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerExpression.R
\name{enhancerExpression}
\alias{enhancerExpression}
\title{Function which associated at each enhancer the gene expression level}
\usage{
enhancerExpression(enhancerTable, geneExpressionTable)
}
\arguments{
\item{enhancerTable}{GRanges table contains genomic position (ex. return by ChromHMM tool)}

\item{geneExpressionTable}{a table contains gene expression and gene Ensembl name (RNAseq data)}
}
\value{
GRanges table with for each genomic position all gene associated and the expression of these genes
}
\description{
Function which associated at each enhancer the gene expression level
}
