% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnvironment.R
\name{geneEnvironment}
\alias{geneEnvironment}
\title{Function with associate at each enhancer the nearly gene}
\usage{
geneEnvironment(table, table_chromHMM, state_order, interval = 3000)
}
\arguments{
\item{table}{a bed genome annotation file}

\item{table_chromHMM}{a ChromHMM table}

\item{state_order}{a list of state}

\item{interval}{a numeric value corresponding to environment distance (default = 3000)}
}
\value{
the table with the distribution of each chromatin state in the environment
}
\description{
Function with associate at each enhancer the nearly gene
}
