% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_file.R
\docType{data}
\name{genome_file}
\alias{genome_file}
\title{BED file contains name, position of genes in mm10 genome}
\format{
A data frame with 54446 rows and 6 variables:
\describe{
\item{gene_ENS}{gene name in ENSEMBL nomenclature}
\item{chr}{chromosome number}
\item{start}{start position, in bp}
\item{end}{end position, in bp}
\item{strand}{strand}
\item{score}{score of enrichent}
}
}
\usage{
genome_file
}
\description{
BED file contains name, position of genes in mm10 genome
}
\keyword{datasets}
