% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomeFile.R
\docType{data}
\name{genomeFile}
\alias{genomeFile}
\title{BED file contains name, position of genes in mm10 genome}
\format{
A GRanges object with 54446 ranges and 2 metadata columns:
\describe{
\item{gene_ENS}{gene name in ENSEMBL nomenclature}
\item{chr}{chromosome number}
\item{start}{start position, in bp}
\item{end}{end position, in bp}
\item{strand}{strand}
\item{score}{score of enrichent}
}
}
\usage{
genomeFile
}
\description{
BED file contains name, position of genes in mm10 genome
}
\keyword{datasets}
