% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInformation.R
\name{getInformation}
\alias{getInformation}
\title{getInformation}
\usage{
getInformation(dataTable)
}
\arguments{
\item{dataTable}{GRanges object or list of GRanges output of enhancerExpression}
}
\value{
distance, expression, gene name information
}
\description{
Function to return information (expression and distance) for each enhancer
}
