% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStateColor.R
\name{getStateColor}
\alias{getStateColor}
\title{getStateColor}
\usage{
getStateColor(colorTable)
}
\arguments{
\item{colorTable}{dataframe color information}
}
\value{
list of vector with each chromatin state is associated to color
}
\description{
Function to return chromatin state color table
}
\examples{
colorTable = system.file("extdata", colorTable, package = "ChromENVEE")
data(colorTable)
colorTable

}
