% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lengthVector.R
\name{lengthVector}
\alias{lengthVector}
\title{Function to create a vector of limit value}
\usage{
lengthVector(lim)
}
\arguments{
\item{lim}{a limit value for distance}
}
\value{
a vector of limit
}
\description{
Function to create a vector of limit value
}
