% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_table_enhancer.R
\docType{data}
\name{list_table_enhancer}
\alias{list_table_enhancer}
\title{A list of 8 GRanges data frame, each data frame contains enhancer chromatin state information}
\format{
A list of 8 GRanges data frame:
\describe{
\item{seqnames}{chromosome number}
\item{ranges}{start and end position, in bp}
\item{strand}{strand value}
\item{chromatin_state}{chromatin state number}
}
}
\usage{
list_table_enhancer
}
\description{
A list of 8 GRanges data frame, each data frame contains enhancer chromatin state information
}
\keyword{datasets}
