% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listTableEnhancer.R
\docType{data}
\name{listTableEnhancer}
\alias{listTableEnhancer}
\title{A list of 8 GRanges data frame, each data frame contains enhancer chromatin state information}
\format{
A list of 8 GRanges data frame:
\describe{
\item{seqnames}{chromosome number}
\item{ranges}{start and end position, in bp}
\item{strand}{strand value}
\item{chromatin_state}{chromatin state number}
\item{sample_name}{name of the sample}
}
}
\usage{
listTableEnhancer
}
\description{
A list of 8 GRanges data frame, each data frame contains enhancer chromatin state information
}
\keyword{datasets}
