% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChromatinState.R
\name{plotChromatinState}
\alias{plotChromatinState}
\title{Function to return summary table from ChromHMM data}
\usage{
plotChromatinState(
  genomic_region,
  state_name,
  plot = T,
  merge = F,
  color = F,
  color_value = "",
  filename = "chromatin_state_distribution",
  ylab = "number chromatin state (\%)",
  xlab = ""
)
}
\arguments{
\item{genomic_region}{a table contains genomic region associate each chromatin state (output from ChromHH)}

\item{state_name}{a vector of chromatin state name (name from ChromHMM)}
}
\value{
the table contains distribution different chromatin state
}
\description{
Function to return summary table from ChromHMM data
}
