% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnhancerExpression.R
\name{plotEnhancerExpression}
\alias{plotEnhancerExpression}
\title{plotEnhancerExpression}
\usage{
plotEnhancerExpression(
  dataTable,
  xlab = "",
  ylab = "gene expression (CPM)",
  scale = "none",
  colorTable,
  distance = 0
)
}
\arguments{
\item{dataTable}{GRanges object or list of GRanges output of enhancerExpression}

\item{xlab}{x-axis label (default = "")}

\item{ylab}{y-axis label (default = "gene expression (CPM)")}

\item{scale}{value (log10, log2 or none) to rescale expression (default = "none")}

\item{colorTable}{dataframe with color information}

\item{distance}{numeric value to focus on region (default = 0)}
}
\value{
ggplot2 figure corresponding to the distribution of gene expression
}
\description{
Function to create a plot which represent the distribution of gene expression
}
\examples{
data(listTableEnhancer)
data(genomeFile)
data(geneExpression)
data(colorTable)
anno = enhancerAnnotation(listTableEnhancer[[1]], genomeFile)
expression = enhancerExpression(anno, geneExpression)
plotEnhancerExpression(expression, colorTable = colorTable)

}
