% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneDistance.R
\name{plotGeneDistance}
\alias{plotGeneDistance}
\title{Function with associate at each enhancer the nearly gene}
\usage{
plotGeneDistance(data_table, xlab = "", ylab = "distance enhancer-gene (bp)")
}
\arguments{
\item{enhancer_table}{GRanges table contains genomic position return by ChromHMM tool}

\item{genome}{a bed genome annotation file}

\item{interval}{a numeric value corresponding to the distance from enhancer where gene is looking for}
}
\value{
the table with infromation concerning gene associate enhancer
}
\description{
Function with associate at each enhancer the nearly gene
}
