% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predominantState.R
\name{predominantState}
\alias{predominantState}
\title{predominantState}
\usage{
predominantState(
  table,
  state,
  header,
  neighbors = 32,
  metric = "euclidean",
  dist = 0.5
)
}
\arguments{
\item{table}{GRanges with the output of geneEnvironment}

\item{state}{list of chromatin state}

\item{header}{list of column in table to analysed with UMAP}

\item{neighbors}{numeric value (see umap package)}

\item{metric}{chracter value (see umap package)}

\item{dist}{numeric value (see umap package)}
}
\value{
table with 2-dimensional value and predominant chromatin state
}
\description{
Function with estimate the predominant state associated at each gene
}
\examples{
geneFile = system.file("extdata", "gene.tsv", package = "ChromENVEE")
geneExpression = read.table(geneFile, h = TRUE)
data(chromatinState)
data(colorTable)
enviro = geneEnvironment(geneExpression, chromatinState, unique(colorTable$stateName))
predominantState(enviro, state = unique(colorTable$stateName),
  header = unique(colorTable$stateName))

}
