% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themePlot.R
\name{themePlot}
\alias{themePlot}
\title{Function to return the theme usable for each plot}
\usage{
themePlot()
}
\value{
the theme usable with ggplot
}
\description{
Function to return the theme usable for each plot
}
