% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorTable.R
\docType{data}
\name{colorTable}
\alias{colorTable}
\title{colorTable}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{stateNumber}{chromatin state number}
\item{stateName}{chromatin state name}
\item{colorValue}{chromatin state color}
}
}
\usage{
colorTable
}
\description{
Data frame contains color and state name information
}
\keyword{datasets}
