% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerAnnotation.R
\name{enhancerAnnotation}
\alias{enhancerAnnotation}
\title{enhancerAnnotation}
\usage{
enhancerAnnotation(enhancerTable, genome, interval = 5e+05, nCore = 1)
}
\arguments{
\item{enhancerTable}{GRanges object or list of GRanges object with enhancer position}

\item{genome}{GRanges of genome annotation file}

\item{interval}{distance from enhancer where gene is looking for (default = 500000 (500kb))}

\item{nCore}{number of core used (default nCore = 1)}
}
\value{
GRanges with infromation about gene associated with enhancer
}
\description{
Function which associated at each enhancer all gene in the interval arount the enhancer
}
\examples{
data(listTableEnhancer)
data(genomeFile)
anno = enhancerAnnotation(listTableEnhancer[[1]], genomeFile)
anno

}
