% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enhancerExpression.R
\name{enhancerExpression}
\alias{enhancerExpression}
\title{enhancerExpression}
\usage{
enhancerExpression(enhancerTable, geneExpressionTable)
}
\arguments{
\item{enhancerTable}{GRanges object contains genomic position}

\item{geneExpressionTable}{gene expression and gene Ensembl name (RNAseq data)}
}
\value{
GRanges table with for each genomic position all gene associated and the expression of these genes
}
\description{
Function which associated at each enhancer the gene expression level
}
\examples{
data(listTableEnhancer)
data(genomeFile)
data(geneExpression)
anno = enhancerAnnotation(listTableEnhancer[[1]], genomeFile)
enhancerExpression(anno, geneExpression)

}
