% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneEnvironment.R
\name{geneEnvironment}
\alias{geneEnvironment}
\title{Function which estimated the chromatin state in the environment around gene (estimated with interval around gene TSS)}
\usage{
geneEnvironment(
  geneExpressionTable,
  tableChromatinState,
  stateOrder,
  interval = 3000
)
}
\arguments{
\item{geneExpressionTable}{a table contains gene expression and gene Ensembl name (ex. geneExpression data)}

\item{tableChromatinState}{a bed table containing information about the chromatin state (ex. chromatin_state data)}

\item{stateOrder}{a list of chromatin state}

\item{interval}{a numeric value corresponding to environment distance (default = 3000 (3kb))}
}
\value{
table with the distribution of each chromatin state in the environment
}
\description{
Function which estimated the chromatin state in the environment around gene (estimated with interval around gene TSS)
}
