% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneExpression.R
\name{geneExpression}
\alias{geneExpression}
\title{Function with associate at each enhancer the gene expression level}
\usage{
geneExpression(enhancer_table, gene_expression_table)
}
\arguments{
\item{enhancer_table}{GRanges table contains genomic position return by ChromHMM tool}

\item{gene_expression_table}{a table contains gene expression and gene Ensembl gene (RNAseq)}
}
\value{
GRanges table with for each genomic position gene associate and the expression of gene
}
\description{
Function with associate at each enhancer the gene expression level
}
