% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInformation.R
\name{getInformation}
\alias{getInformation}
\title{Function to return information (expression and distance) for each enhancer}
\usage{
getInformation(dataTable)
}
\arguments{
\item{dataTable}{a GRanges table or list of GRanges obtains by enhancerExpression function}
}
\value{
the ggplot2 figure corresponding to the distribution of distance gene-enhancer
}
\description{
Function to return information (expression and distance) for each enhancer
}
