% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLengthVector.R
\name{getLengthVector}
\alias{getLengthVector}
\title{getLengthVector}
\usage{
getLengthVector(lim)
}
\arguments{
\item{lim}{limit value for distance}
}
\value{
a vector of limit
}
\description{
Function to create a vector of limit value
}
