% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStateColor.R
\name{getStateColor}
\alias{getStateColor}
\title{Function to return chromatin state color table}
\usage{
getStateColor(state_name, state_number, color)
}
\arguments{
\item{state_name}{a vector of chromatin state name}

\item{state_number}{a vector of chromatin state number}

\item{color}{a vector of color to associated to each chromatin state}
}
\value{
list of vector with each chromatin state associate to color
}
\description{
Function to return chromatin state color table
}
