% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStateColor.R
\name{getStateColor}
\alias{getStateColor}
\title{Function to return chromatin state color table}
\usage{
getStateColor(colorTable)
}
\arguments{
\item{colorTable}{a data frame which contains color information}
}
\value{
list of vector with each chromatin state is associated to color
}
\description{
Function to return chromatin state color table
}
