% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChromatinState.R
\name{plotChromatinState}
\alias{plotChromatinState}
\title{Function to return summary table from ChromHMM data and create associated plot}
\usage{
plotChromatinState(
  tableChromatinState,
  colorTable,
  plot = T,
  merge = F,
  filename = "chromatin_state_distribution",
  ylab = "chromatin state (\%)",
  xlab = ""
)
}
\arguments{
\item{tableChromatinState}{a bed table containing information about the chromatin state (ex. chromatin_state data)}

\item{colorTable}{a data frame which contains color information}

\item{plot}{a boolean to create plot (default = T)}

\item{merge}{a boolean to merge data if it's a list of table (default = F), if TRUE, list of dataframe is merge}

\item{filename}{a string to name the plot create (defualt = "chromatin_state_distribution")}

\item{ylab}{a string (default = "chromatin state (\%)")}

\item{xlab}{a string (default = "")}
}
\value{
table contains distribution of different chromatin state
}
\description{
Function to return summary table from ChromHMM data and create associated plot
}
