% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChromatinState.R
\name{plotChromatinState}
\alias{plotChromatinState}
\title{plotChromatinState}
\usage{
plotChromatinState(
  tableChromatinState,
  colorTable,
  plot = TRUE,
  merge = FALSE,
  filename = "chromatin_state_distribution",
  ylab = "chromatin state (\%)",
  xlab = ""
)
}
\arguments{
\item{tableChromatinState}{GRanges object containing information about the chromatin state (ex. chromatinState data)}

\item{colorTable}{dataframe which contains color information (ex. colorTable data)}

\item{plot}{boolean to create plot (default = TRUE)}

\item{merge}{boolean to merge data if it's a list of table (default = F), if TRUE, list of dataframe is merge}

\item{filename}{string to name the plot create (default = "chromatin_state_distribution")}

\item{ylab}{x-axis label (default = "chromatin state (\%)")}

\item{xlab}{y-axis label (default = "")}
}
\value{
table contains distribution of different chromatin state
}
\description{
Function to return summary table from ChromHMM data and create associated plot
}
\examples{
chromatinState = system.file("extdata", chromatinState, package = "ChromENVEE")
data(chromatinState)
colorTable = system.file("extdata", colorTable, package = "ChromENVEE")
data(colorTable)
state = plotChromatinState(chromatinState, colorTable)
state

}
