% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistanceExpression.R
\name{plotDistanceExpression}
\alias{plotDistanceExpression}
\title{Function to create a plot which represent the gene expression in function distance gene-enhancer}
\usage{
plotDistanceExpression(
  dataTable,
  xlab = "",
  ylab = "log(CPM)",
  colorTable,
  limit = 5e+05
)
}
\arguments{
\item{dataTable}{a GRanges table or list of GRanges obtains by enhancerExpression function}

\item{xlab}{a string (default = "")}

\item{ylab}{a string (default = "log(CPM)")}

\item{colorTable}{a data frame which contains color information}

\item{limit}{a value of limit for distance analysis (default = 500000 (500kb))}
}
\value{
ggplot2 figure corresponding to the mean of gene expression in function distance gene-enhancer
}
\description{
Function to create a plot which represent the gene expression in function distance gene-enhancer
}
