% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistanceExpression.R
\name{plotDistanceExpression}
\alias{plotDistanceExpression}
\title{plotDistanceExpression}
\usage{
plotDistanceExpression(
  dataTable,
  xlab = "",
  ylab = "log(CPM)",
  colorTable,
  limit = 5e+05
)
}
\arguments{
\item{dataTable}{GRanges object or list of GRanges output of enhancerExpression}

\item{xlab}{x-axis label (default = "")}

\item{ylab}{y-axis label (default = "log(CPM)")}

\item{colorTable}{dataframe which contains color information (colorTable data)}

\item{limit}{limit for distance analysis (default = 500000 (500kb))}
}
\value{
gene expression in function distance gene-enhancer
}
\description{
Function to create a plot which represent the gene expression in function distance gene-enhancer
}
\examples{
listTableEnhancer = system.file("extdata", listTableEnhancer, package = "ChromENVEE")
data(listTableEnhancer)
genomeFile = system.file("extdata", genomeFile, package = "ChromENVEE")
data(genomeFile)
geneExpression = system.file("extdata", geneExpression, package = "ChromENVEE")
data(geneExpression)
colorTable = system.file("extdata", "colorTable", package = "ChromENVEE")
data(colorTable)
anno = enhancerAnnotation(listTableEnhancer[[1]], genomeFile)
expression = enhancerExpression(anno, geneExpression)
plotDistanceExpression(expression, colorTable = colorTable)

}
