% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistanceExpression.R
\name{plotDistanceExpression}
\alias{plotDistanceExpression}
\title{Function to create a plot which represent the gene expression in function distance gene-enhancer}
\usage{
plotDistanceExpression(
  data_table,
  xlab = "",
  ylab = "log(CPM)",
  color,
  state_number,
  state_name
)
}
\arguments{
\item{data_table}{a GRanges table or list of GRanges obtains by enhancerExpression function}

\item{xlab}{a string}

\item{ylab}{a string}

\item{color}{a list of color value}

\item{state_number}{a list of chromatin state number}

\item{state_name}{a list of chromatin state name}
}
\value{
the ggplot2 figure corresponding to the mean of gene expression in function distance gene-enhancer
}
\description{
Function to create a plot which represent the gene expression in function distance gene-enhancer
}
