% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnhancerExpression.R
\name{plotEnhancerExpression}
\alias{plotEnhancerExpression}
\title{Function to create a plot which represent the distribution of gene expression}
\usage{
plotEnhancerExpression(
  data_table,
  xlab = "",
  ylab = "gene expression log10(cpm)",
  scale = "none",
  color,
  state_number,
  state_name,
  distance = 0
)
}
\arguments{
\item{data_table}{a GRanges table or list of GRanges obtains by enhancerExpression function}

\item{xlab}{a string}

\item{ylab}{a string}

\item{scale}{a value (log10, log2 or none) to rescale expression (default = "none")}

\item{color}{a list of color value}

\item{state_number}{a list of chromatin state number}

\item{state_name}{a list of chromatin state name}

\item{distance}{a value to zoom the expression focus (default = 0)}
}
\value{
the ggplot2 figure corresponding to the distribution of gene expression
}
\description{
Function to create a plot which represent the distribution of gene expression
}
