% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEnhancerExpression.R
\name{plotEnhancerExpression}
\alias{plotEnhancerExpression}
\title{Function to create a plot which represent the distribution of gene expression}
\usage{
plotEnhancerExpression(
  dataTable,
  xlab = "",
  ylab = "gene expression (CPM)",
  scale = "none",
  color,
  stateNumber,
  stateName,
  distance = 0
)
}
\arguments{
\item{dataTable}{a GRanges table or list of GRanges obtains by enhancerExpression function}

\item{xlab}{a string (default = "")}

\item{ylab}{a string (default = "gene expression (CPM)")}

\item{scale}{a value (log10, log2 or none) to rescale expression (default = "none")}

\item{color}{a list of color value}

\item{stateNumber}{a list of chromatin state number}

\item{stateName}{a list of chromatin state name}

\item{distance}{a value to zoom the expression focus (default = 0)}
}
\value{
ggplot2 figure corresponding to the distribution of gene expression
}
\description{
Function to create a plot which represent the distribution of gene expression
}
