% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneAssociation.R
\name{plotGeneAssociation}
\alias{plotGeneAssociation}
\title{plotGeneAssociation}
\usage{
plotGeneAssociation(table, all = FALSE)
}
\arguments{
\item{table}{GRanges object or list of GRanges output of enhancerAnnotation}

\item{all}{boolean, if TRUE, list of dataframe is merge (default = FALSE)}
}
\value{
ggplot2 figure of the number of gene associated to enhancer
}
\description{
Function which plot the number of gene associated to enhancer
}
\examples{
data(listTableEnhancer)
data(genomeFile)
anno = enhancerAnnotation(listTableEnhancer[[1]], genomeFile)
plotGeneAssociation(anno)

}
