% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneAssociation.R
\name{plotGeneAssociation}
\alias{plotGeneAssociation}
\title{Function which plot the number of gene associated to enhancer}
\usage{
plotGeneAssociation(table, all = F)
}
\arguments{
\item{table}{a GRanges table or list of GRanges obtains by enhancerAnnotation function}

\item{all}{a boolean, if TRUE, list of dataframe is merge (default = F)}
}
\value{
ggplot2 figure of the number of gene associated to enhancer
}
\description{
Function which plot the number of gene associated to enhancer
}
