% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneAssociation.R
\name{plotGeneAssociation}
\alias{plotGeneAssociation}
\title{Function which plot the number of gene associated to enhancer}
\usage{
plotGeneAssociation(table, all = F)
}
\arguments{
\item{table}{GRanges table contains results obtains by enhancerAnnotation()}

\item{all}{a boolean, if TRUE, list of dataframe is merge}
}
\value{
a plot of the number of gene associated to enhancer
}
\description{
Function which plot the number of gene associated to enhancer
}
