% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneDistance.R
\name{plotGeneDistance}
\alias{plotGeneDistance}
\title{Function to create a plot which represent the distribution of distance gene-enhancer}
\usage{
plotGeneDistance(
  dataTable,
  limit = 5e+05,
  xlab = "",
  ylab = "distance enhancer-gene (bp)"
)
}
\arguments{
\item{dataTable}{a GRanges table or list of GRanges obtains by enhancerExpression function}

\item{limit}{a value of limit for distance analysis (default = 500000 (500kb))}

\item{xlab}{a string (default = "")}

\item{ylab}{a string (default = "distance enhancer-gene (bp))}
}
\value{
ggplot2 figure corresponding to the distribution of distance gene-enhancer
}
\description{
Function to create a plot which represent the distribution of distance gene-enhancer
}
