% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGeneDistance.R
\name{plotGeneDistance}
\alias{plotGeneDistance}
\title{plotGeneDistance}
\usage{
plotGeneDistance(
  dataTable,
  limit = 5e+05,
  xlab = "",
  ylab = "distance enhancer-gene (bp)"
)
}
\arguments{
\item{dataTable}{GRanges object or list of GRanges output of enhancerExpression}

\item{limit}{limit for distance analysis (default = 500000 (500kb))}

\item{xlab}{x-axis label (default = "")}

\item{ylab}{y-axis label (default = "distance enhancer-gene (bp))}
}
\value{
distribution of distance gene-enhancer
}
\description{
Function to create a plot which represent the distribution of distance gene-enhancer
}
\examples{
listTableEnhancer = system.file("extdata", listTableEnhancer, package = "ChromENVEE")
data(listTableEnhancer)
genomeFile = system.file("extdata", genomeFile, package = "ChromENVEE")
data(genomeFile)
geneExpression = system.file("extdata", geneExpression, package = "ChromENVEE")
data(geneExpression)
anno = enhancerAnnotation(listTableEnhancer[[1]], genomeFile)
expression = enhancerExpression(anno, geneExpression)
plotGeneDistance(expression)

}
