% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predominantState.R
\name{predominantState}
\alias{predominantState}
\title{Function with estimate the predominant state associated at each gene}
\usage{
predominantState(
  table,
  state,
  header,
  neighbors = 32,
  metric = "euclidean",
  dist = 0.5
)
}
\arguments{
\item{table}{a table corresponding to the result of geneEnvironment function}

\item{state}{a list of chromatin state}

\item{header}{a list of column in table to analysed with UMAP}

\item{neighbors}{a numeric value (see umapr package)}

\item{metric}{a chracter value (see umapr package)}

\item{dist}{a numeric value (see umapr package)}
}
\value{
table with 2-dimensional value and predominant chromatin state
}
\description{
Function with estimate the predominant state associated at each gene
}
