% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predominentState.R
\name{predominentState}
\alias{predominentState}
\title{Function with associate at each enhancer the nearly gene}
\usage{
predominentState(
  table,
  state,
  header,
  neighbors = 32,
  metric = "euclidean",
  dist = 0.5
)
}
\arguments{
\item{table}{a table corresponding to the result of geneEnvironment function}

\item{state}{a list of chromatin state}

\item{header}{a list of column in table to analysed with UMAP}

\item{metric}{a chracter value}

\item{dist}{a numeric value}

\item{neighbor}{a numeric value}
}
\value{
the table with 2-dimensional value and predominent chromatin state
}
\description{
Function with associate at each enhancer the nearly gene
}
